Split navbar file in several files
