+++ src/theme/components/material-kit/Navbar/index.ts
@@ -28,4 +28,4 @@
 
  * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
  */
-export * from './Navbar';
+export { Navbar } from './Navbar';
